@@ -14,15 +14,19 @@
     <name>SpringBootSoap</name>
     <description>Spring Boot Soap Client</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <maven.plugin.validation>verbose</maven.plugin.validation>
     </properties>
+
     <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web-services</artifactId>
         </dependency>
-
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-devtools</artifactId>
